renamed a few files to make it clearer what is singleSPA related and what not
