+++ app1React/src/singleSpaEntry.js
