+++ app2Angular4/src/main-module.ts
@@ -7,7 +7,7 @@ import {enableProdMode} from '@angular/core';
 import {RouterModule, Routes} from "@angular/router";
 import {APP_BASE_HREF} from "@angular/common";
 import { NgReduxModule, NgRedux } from '@angular-redux/store';
-import { IAppState, CounterActions } from './app2Store';
+import { IAppState, CounterActions } from './store';
 import { Globals } from "./globals.service";
 
 const appRoutes: Routes = [
