+++ app2Angular4/src/singleSpaEntry.js
