+++ app2Angular4/src/store.ts
