+++ app3Angular1/src/singleSpaEntry.js
