+++ portal/src/portal.js
@@ -6,13 +6,13 @@ async function init() {
     const globalEventDistributor = new GlobalEventDistributor();
 
     // app1: The URL "/app1/..." is being rewritten to "http://localhost:9001/..." this is done by the webpack proxy
-    await loadApp('app1', '/app1', '/app1/main.js', '/app1/store.js', globalEventDistributor);
+    await loadApp('app1', '/app1', '/app1/singleSpaEntry.js', '/app1/store.js', globalEventDistributor);
 
     // app2: The URL "/app2/..." is being rewritten to "http://localhost:9002/..." this is done by the webpack proxy
-    await loadApp('app2', '/app2', '/app2/main.js', '/app2/store.js', globalEventDistributor);
+    await loadApp('app2', '/app2', '/app2/singleSpaEntry.js', '/app2/store.js', globalEventDistributor);
 
     // app3: The URL "/app3/..." is being rewritten to "http://localhost:9003/..." this is done by the webpack proxy
-    await loadApp('app3', '/app3', '/app3/main.js', null, null); // does not have a store, so we pass null
+    await loadApp('app3', '/app3', '/app3/singleSpaEntry.js', null, null); // does not have a store, so we pass null
 
     singleSpa.start();
 }
